enabled the track order button

+++ src/utils/index.ts
@@ -26,6 +26,6 @@ export async function verifyLink(url: string){
 }
 
 export const formatNumber = (number: number | string) => {
-  number = typeof number === 'number' ? number : parseInt(number)
+  number = typeof number === 'number' ? number : parseFloat(number)
   return number.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",");
 };
